refactor: absorb thirdparty packages into pkg and remove thirdparty folder (#4554)
