+++ .github/copilot-code-review.yml
@@ -47,8 +47,6 @@ review:
       instructions: "Do not review — generated by k8s code-generator"
     - path: "**/zz_generated*"
       instructions: "Do not review — generated by controller-tools"
-    - path: "thirdparty/**"
-      instructions: "Do not review — vendored upstream code"
     - path: "go.sum"
       instructions: "Do not review"
     - path: "go.mod"
