+++ .golangci.yml
@@ -66,7 +66,6 @@ linters:
           - gosec
           - funlen
     paths:
-      - thirdparty/
       - third_party$
       - builtin$
       - examples$
@@ -77,7 +76,6 @@ formatters:
   exclusions:
     generated: lax
     paths:
-      - thirdparty/
       - third_party$
       - builtin$
       - examples$
