+++ AGENTS.md
@@ -173,7 +173,6 @@ git config --global user.name "Your Name"
 * **release/**: Release automation (GoReleaser config, Homebrew formula generation)
 * **hack/**: Miscellaneous development utilities
 * **healthcheck/**: Separate module for health checking (Go (Go version is defined in [healthcheck/go.mod](./healthcheck/go.mod)), local Makefile)
-* **thirdparty/**: Third-party code (excluded from linting)
 * **Formula/**: Homebrew package definition (generated by `go run ./release/formula/main.go VERSION`)
 
 ## Linting Rules & Style
