+++ pkg/cmdconfig/commands/cmdcat/cmdcat.go
@@ -27,8 +27,8 @@ import (
 
 	kptfilev1 "github.com/kptdev/kpt/api/kptfile/v1"
 	"github.com/kptdev/kpt/internal/docs/generated/pkgdocs"
+	"github.com/kptdev/kpt/pkg/cmdconfig/commands/runner"
 	argsutil "github.com/kptdev/kpt/pkg/lib/util/args"
-	"github.com/kptdev/kpt/thirdparty/cmdconfig/commands/runner"
 	"github.com/spf13/cobra"
 	"sigs.k8s.io/kustomize/kyaml/kio"
 	"sigs.k8s.io/kustomize/kyaml/kio/filters"
