+++ pkg/cmdconfig/commands/cmdsink/cmdsink.go
