+++ pkg/cmdconfig/commands/cmdsink/cmdsink_test.go
