+++ pkg/fn/runfn/test/testdata/java/java-configmap.resource.yaml
