+++ pkg/fn/runfn/test/testdata/java/java-deployment.resource.yaml
