+++ pkg/fn/runfn/test/testdata/java/java-service.resource.yaml
