+++ sonar.properties
@@ -6,11 +6,11 @@ sonar.organization=kptdev
 sonar.language=go
 
 # Path to your Go source code
-# Includes all relevant source directories, excluding vendor and thirdparty
+# Includes all relevant source directories, excluding vendor
 sonar.sources=commands,pkg,run,api/fnresult,api/kptfile,api/resourcegroup,api/schema
 
 # Exclude files if needed
-sonar.exclusions=**/test/**, **/examples/*, **/scripts/*, **/*_test.go, **/testing*, **/generated/**, **/testdata/**, **/*zz_generated.*, vendor/**, thirdparty/**, .github/**, documentation/**, Formula/**
+sonar.exclusions=**/test/**, **/examples/*, **/scripts/*, **/*_test.go, **/testing*, **/generated/**, **/testdata/**, **/*zz_generated.*, vendor/**, .github/**, documentation/**, Formula/**
 
 # To include test coverage reports
 sonar.test.inclusions=**/*_test.go
